requirements.txt
flask
python-telegram-bot
python-dotenv
apscheduler
firebase-admin